+++ docs/changelog.rst
@@ -11,10 +11,12 @@ for Python 3.6).
 
 New features:
 
+* Added option to specify a URL in the `embedart` plugin.
+  :bug:`83`
 * :ref:`list-cmd` `singleton:true` queries have been made faster
 * :ref:`list-cmd` `singleton:1` and `singleton:0` can now alternatively be used in queries, same as `comp`
 * --from-logfile now parses log files using a UTF-8 encoding in `beets/beets/ui/commands.py`.
-  :bug:`4693` 
+  :bug:`4693`
 * :doc:`/plugins/bareasc` lookups have been made faster
 * :ref:`list-cmd` lookups using the pattern operator `::` have been made faster
 * Added additional error handling for `spotify` plugin.